+++ src/api/articles.ts
@@ -1,4 +1,4 @@
-import axiosInstance from './axiosInstance';
+import axiosInstance from '../app/api/axiosInstance';
 
 /* 자유게시판 - 일반 글 */
 export const fetchGeneral = (keyword?: string) => {
